+++ src/MtgCollectionTracker/Services/Stubs/StubCollectionTrackingService.cs
@@ -296,7 +296,7 @@ public class StubCollectionTrackingService : ICollectionTrackingService
         throw new System.NotImplementedException();
     }
 
-    public ValueTask<(decimal? price, string? provider)> GetLatestPriceForSkuAsync(Guid skuId, string currency, CancellationToken cancel)
+    public ValueTask<(double? price, string? provider)> GetLatestPriceForSkuAsync(Guid skuId, string currency, CancellationToken cancel)
     {
         throw new System.NotImplementedException();
     }
